+++ CHANGELOG.md
@@ -9,6 +9,10 @@ All notable changes to this project will be documented in this file.
 
 - New `cyborgdb` output. (@ahellegit)
 
+### Changed
+
+- Fixed an issue where MCP output tools would yield invalid JSON Schema properties. (@Jeffail)
+
 ## 4.65.0 - 2025-09-23
 
 ### Added